Split chain_spec to individual chain families.
